misha 4.0 OSX support Bug fix in all functions using 2D intervals set iterator: in multitasking mode some of the chromosome pairs might be skipped. In non-multitasking mode the scope might be considered as empty. This behavior is random and recurrent calls might suddenly return correct results. Bug fix in multitasking: occasional hang ups when memory usage of the child processes exceeds the limit gmax.mem.usage Bug fix in multitasking: all functions creating new files and reporting progress might create corrupted files \(gtrack.create, ....\) Bug fix in all functions using intervals.set.out parameter: small intervals set instead of big one might be created and vice versa. Also an error message 'result size exceeded the maximum allowed' might be mistakenly generated  Bug fix in gintervals.load applied to 1D big intervals: 'Error in if (progress && percentage < 100 && progress.percentage != percentage)' Bug fix in all functions returning 1D or 2D intervals: in rare random cases NULL or invalid intervals set is returned Bug fix in gcompute_strands_autocorr: internal buffer overflow and possible memory corruption gdb.reload: run-time improvements

+++ src/GenomeTrackWilcox.cpp
@@ -153,7 +153,7 @@ void GenomeTrackSlidingWilcox::set_next_sample(double v)
 
 	if (std::isnan(m_queue[m_center]) || z > m_maxz) {
 		if (m_start_coord != -1) {
-			int64_t start = max(0L, m_start_coord - m_winsize_aside[SMALL] * m_binsize);
+			int64_t start = max((int64_t)0, m_start_coord - m_winsize_aside[SMALL] * m_binsize);
 			int64_t end = m_center_coord + m_winsize_aside[SMALL] * m_binsize;
 
 			if (m_intervals.empty() || m_intervals.back().chromid != m_chromid || m_intervals.back().end < start)
